Rolled back test changes to get build working
