+++ plugin/cds-hooks/pom.xml
@@ -24,12 +24,12 @@
 			<artifactId>cqf-ruler-security</artifactId>
 			<version>0.5.1-SNAPSHOT</version>
 		</dependency>
-        <dependency>
-            <groupId>org.opencds.cqf.ruler</groupId>
-            <artifactId>cqf-ruler-cpg</artifactId>
-            <version>0.5.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
+		<dependency>
+			<groupId>org.opencds.cqf.ruler</groupId>
+			<artifactId>cqf-ruler-cpg</artifactId>
+			<version>0.5.1-SNAPSHOT</version>
+			<scope>compile</scope>
+		</dependency>
 
-    </dependencies>
+	</dependencies>
 </project>
